+++ contrib/opencensus-ext-azure/examples/traces/simple.py
@@ -23,3 +23,4 @@ tracer = Tracer(exporter=AzureExporter(), sampler=ProbabilitySampler(1.0))
 
 with tracer.span(name='foo'):
     print('Hello, World!')
+input(...)