+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Attach rate metrics for VM
   ([#935](https://github.com/census-instrumentation/opencensus-python/pull/935))
 - Add links in properties for trace exporter envelopes
-  ([#935](https://github.com/census-instrumentation/opencensus-python/pull/935))
+  ([#936](https://github.com/census-instrumentation/opencensus-python/pull/936))
 
 ## 1.0.4
 Released 2020-06-29